refactor: split inventory protocol module
